requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-multipart==0.0.6
Pillow==10.0.0
numpy==1.24.3
faiss-cpu==1.7.4
pymongo==4.4.0
python-dotenv==1.0.0